+++ src/core/createClientWorld.js
@@ -8,7 +8,7 @@ import { ClientGraphics } from './systems/ClientGraphics'
 import { ClientEnvironment } from './systems/ClientEnvironment'
 import { ClientStats } from './systems/ClientStats'
 import { ClientEditor } from './systems/ClientEditor'
-import { Chat } from './systems/Chat'
+import { ClientActions } from './systems/ClientActions'
 
 export function createClientWorld() {
   const world = new World()
@@ -18,7 +18,8 @@ export function createClientWorld() {
   world.register('loader', ClientLoader)
   world.register('graphics', ClientGraphics)
   world.register('environment', ClientEnvironment)
-  // world.register('stats', ClientStats)
+  world.register('stats', ClientStats)
   world.register('editor', ClientEditor)
+  world.register('actions', ClientActions)
   return world
 }