+++ src/core/entities/App.js
@@ -1,3 +1,4 @@
+import { isString } from 'lodash-es'
 import * as THREE from '../extras/three'
 
 import { Entity } from './Entity'
@@ -8,6 +9,7 @@ import { LerpQuaternion } from '../extras/LerpQuaternion'
 import { ControlPriorities } from '../extras/ControlPriorities'
 
 const hotEventNames = ['fixedUpdate', 'update', 'lateUpdate']
+const internalEvents = ['fixedUpdate', 'updated', 'lateUpdate']
 
 const Modes = {
   ACTIVE: 'active',
@@ -21,15 +23,28 @@ export class App extends Entity {
     super(world, data, local)
     this.isApp = true
     this.n = 0
+    this.nodes = new Map()
     this.worldNodes = new Set()
     this.hotEvents = 0
+    this.worldListeners = new Map()
+    this.listeners = {}
+    this.eventQueue = []
     this.build()
   }
 
+  createNode(data) {
+    const node = createNode(data)
+    if (this.nodes.has(node.id)) {
+      console.error('node with id already exists: ', node.id)
+      return
+    }
+    this.nodes.set(node.id, node)
+    return node
+  }
+
   async build(crashed) {
+    this.building = true
     const n = ++this.n
-    // create base
-    const base = createNode({ name: 'group' })
     // fetch blueprint
     const blueprint = this.world.blueprints.get(this.data.blueprint)
     // fetch script (if any)
@@ -38,31 +53,31 @@ export class App extends Entity {
       script = this.world.loader.get('script', blueprint.script)
       if (!script) script = await this.world.loader.load('script', blueprint.script)
     }
-    let model
+    let root
     // if someone else is uploading glb, show a loading indicator
     if (this.data.uploader && this.data.uploader !== this.world.network.id) {
-      model = createNode({ name: 'mesh' })
-      model.type = 'box'
-      model.width = 1
-      model.height = 1
-      model.depth = 1
+      root = createNode({ name: 'mesh' })
+      root.type = 'box'
+      root.width = 1
+      root.height = 1
+      root.depth = 1
     }
     // otherwise we can load the actual glb
     else {
       try {
         let glb = this.world.loader.get('glb', blueprint.model)
         if (!glb) glb = await this.world.loader.load('glb', blueprint.model)
-        model = glb.toNodes()
+        root = glb.toNodes()
       } catch (err) {
         console.error(err)
         // no model, will use crash block below
       }
     }
     // if script crashed (or failed to load model), show crash-block
-    if (crashed || !model) {
+    if (crashed || !root) {
       let glb = this.world.loader.get('glb', 'asset://crash-block.glb')
       if (!glb) glb = await this.world.loader.load('glb', 'asset://crash-block.glb')
-      model = glb.toNodes()
+      root = glb.toNodes()
     }
     // if a new build happened while we were fetching, stop here
     if (this.n !== n) return
@@ -72,13 +87,17 @@ export class App extends Entity {
     this.mode = Modes.ACTIVE
     if (this.data.mover) this.mode = Modes.MOVING
     if (this.data.uploader && this.data.uploader !== this.world.network.id) this.mode = Modes.LOADING
-    // activate
+    // setup
     this.blueprint = blueprint
-    this.base = base
-    this.base.position.fromArray(this.data.position)
-    this.base.quaternion.fromArray(this.data.quaternion)
-    this.base.add(model)
-    this.base.activate({ world: this.world, entity: this, physics: !this.data.mover })
+    this.root = root
+    this.root.position.fromArray(this.data.position)
+    this.root.quaternion.fromArray(this.data.quaternion)
+    // collect all nodes
+    this.root.traverse(node => {
+      this.nodes.set(node.id, node)
+    })
+    // activate
+    this.root.activate({ world: this.world, entity: this, physics: !this.data.mover })
     // execute script
     if (this.mode === Modes.ACTIVE && script && !crashed) {
       this.script = script
@@ -96,27 +115,37 @@ export class App extends Entity {
     if (this.data.mover === this.world.network.id) {
       this.lastMoveSendTime = 0
       this.control = this.world.controls.bind({
-        priority: ControlPriorities.APP,
+        priority: ControlPriorities.ENTITY,
         onScroll: () => {
           return true
         },
       })
     }
     // if remote is moving, set up to receive network updates
-    this.networkPos = new LerpVector3(base.position, this.world.networkRate)
-    this.networkQuat = new LerpQuaternion(base.quaternion, this.world.networkRate)
+    this.networkPos = new LerpVector3(root.position, this.world.networkRate)
+    this.networkQuat = new LerpQuaternion(root.quaternion, this.world.networkRate)
+    // execute any events we collected while building
+    while (this.eventQueue.length) {
+      const event = this.eventQueue[0]
+      if (event.version > this.blueprint.version) break // ignore future versions
+      this.eventQueue.shift()
+      this.emit(event.name, event.data, event.clientId)
+    }
+    // finished!
+    this.building = false
   }
 
   unbuild() {
     // deactivate local node
-    this.base?.deactivate()
+    this.root?.deactivate()
     // deactivate world nodes
     for (const node of this.worldNodes) {
       node.deactivate()
     }
+    this.nodes.clear()
     this.worldNodes.clear()
     // clear script event listeners
-    this.events = {}
+    this.clearEventListeners()
     this.hotEvents = 0
     // release control
     if (this.control) {
@@ -146,7 +175,7 @@ export class App extends Entity {
     if (this.data.mover === this.world.network.id) {
       if (this.control.buttons.ShiftLeft) {
         // if shift is down we're raising and lowering the app
-        this.base.position.y -= this.world.controls.pointer.delta.y * delta * 0.5
+        this.root.position.y -= this.world.controls.pointer.delta.y * delta * 0.5
       } else {
         // otherwise move with the cursor
         const position = this.world.controls.pointer.position
@@ -160,10 +189,10 @@ export class App extends Entity {
           break
         }
         if (hit) {
-          this.base.position.copy(hit.point)
+          this.root.position.copy(hit.point)
         }
         // and rotate with the mouse wheel
-        this.base.rotation.y += this.control.scroll.delta * 0.1 * delta
+        this.root.rotation.y += this.control.scroll.delta * 0.1 * delta
       }
 
       // periodically send updates
@@ -171,21 +200,23 @@ export class App extends Entity {
       if (this.lastMoveSendTime > this.world.networkRate) {
         this.world.network.send('entityModified', {
           id: this.data.id,
-          position: this.base.position.toArray(),
-          quaternion: this.base.quaternion.toArray(),
+          position: this.root.position.toArray(),
+          quaternion: this.root.quaternion.toArray(),
         })
         this.lastMoveSendTime = 0
       }
       // if we left clicked, we can place the app
       if (this.control.pressed.MouseLeft) {
         this.data.mover = null
-        this.data.position = this.base.position.toArray()
-        this.data.quaternion = this.base.quaternion.toArray()
+        this.data.position = this.root.position.toArray()
+        this.data.quaternion = this.root.quaternion.toArray()
+        this.data.state = {}
         this.world.network.send('entityModified', {
           id: this.data.id,
           mover: null,
           position: this.data.position,
           quaternion: this.data.quaternion,
+          state: this.data.state,
         })
         this.build()
       }
@@ -248,6 +279,10 @@ export class App extends Entity {
       this.data.quaternion = data.quaternion
       this.networkQuat.pushArray(data.quaternion)
     }
+    if (data.hasOwnProperty('state')) {
+      this.data.state = data.state
+      rebuild = true
+    }
     if (rebuild) {
       this.build()
     }
@@ -269,10 +304,10 @@ export class App extends Entity {
   }
 
   on(name, callback) {
-    if (!this.events[name]) {
-      this.events[name] = new Set()
+    if (!this.listeners[name]) {
+      this.listeners[name] = new Set()
     }
-    this.events[name].add(callback)
+    this.listeners[name].add(callback)
     if (hotEventNames.includes(name)) {
       this.hotEvents++
       this.world.setHot(this, this.hotEvents > 0)
@@ -280,8 +315,8 @@ export class App extends Entity {
   }
 
   off(name, callback) {
-    if (!this.events[name]) return
-    this.events[name].delete(callback)
+    if (!this.listeners[name]) return
+    this.listeners[name].delete(callback)
     if (hotEventNames.includes(name)) {
       this.hotEvents--
       this.world.setHot(this, this.hotEvents > 0)
@@ -289,9 +324,37 @@ export class App extends Entity {
   }
 
   emit(name, a1, a2) {
-    if (!this.events[name]) return
-    for (const callback of this.events[name]) {
+    if (!this.listeners[name]) return
+    for (const callback of this.listeners[name]) {
       callback(a1, a2)
+    }
+  }
+
+  onWorldEvent(name, callback) {
+    this.worldListeners.set(callback, name)
+    this.world.events.on(name, callback)
+  }
+
+  offWorldEvent(name, callback) {
+    this.worldListeners.delete(callback)
+    this.world.events.off(name, callback)
+  }
+
+  clearEventListeners() {
+    // local
+    this.listeners = {}
+    // world
+    for (const [callback, name] of this.worldListeners) {
+      this.world.events.off(name, callback)
+    }
+    this.worldListeners.clear()
+  }
+
+  onEvent(version, name, data, socketId) {
+    if (this.building || version > this.blueprint.version) {
+      this.eventQueue.push({ version, name, data, socketId })
+    } else {
+      this.emit(name, data, socketId)
     }
   }
 
@@ -299,6 +362,15 @@ export class App extends Entity {
     const entity = this
     const world = this.world
     return {
+      get networkId() {
+        return world.network.id
+      },
+      get isServer() {
+        return world.network.isServer
+      },
+      get isClient() {
+        return world.network.isClient
+      },
       add(pNode) {
         const node = entity.nodes.get(pNode.id)
         if (!node) return
@@ -316,6 +388,23 @@ export class App extends Entity {
         entity.worldNodes.delete(node)
         node.deactivate()
       },
+      attach(pNode) {
+        const node = entity.nodes.get(pNode.id)
+        if (!node) return
+        const parent = node.parent
+        if (!parent) return
+        parent.remove(node)
+        node.matrix.copy(node.matrixWorld)
+        node.matrix.decompose(node.position, node.quaternion, node.scale)
+        node.activate({ world, entity, physics: true })
+        entity.worldNodes.add(node)
+      },
+      on(name, callback) {
+        entity.onWorldEvent(name, callback)
+      },
+      off(name, callback) {
+        entity.offWorldEvent(name, callback)
+      },
     }
   }
 
@@ -326,14 +415,31 @@ export class App extends Entity {
       get instanceId() {
         return entity.data.id
       },
+      get version() {
+        return entity.blueprint.version
+      },
+      get state() {
+        return entity.data.state
+      },
+      set state(value) {
+        entity.data.state = value
+      },
       on(name, callback) {
         entity.on(name, callback)
       },
       off(name, callback) {
         entity.off(name, callback)
       },
+      send(name, data, ignoreSocketId) {
+        if (internalEvents.includes(name)) {
+          return console.error(`apps cannot emit internal events (${name})`)
+        }
+        // NOTE: on the client ignoreSocketId is a no-op because it can only send events to the server
+        const event = [entity.data.id, entity.blueprint.version, name, data]
+        world.network.send('entityEvent', event, ignoreSocketId)
+      },
       get(id) {
-        const node = entity.base.get(id)
+        const node = entity.root.get(id)
         if (!node) return null
         return node.getProxy()
       },
@@ -347,27 +453,17 @@ export class App extends Entity {
           return node.getProxy()
         }
       },
-      // getState() {
-      //   return entity.state
-      // },
-      // getStateChanges() {
-      //   return entity._stateChanges
-      // },
-      // createNetworkProp(value, onChange) {
-      //   const key = `__${entity.scriptVarIds++}`
-      //   return entity.createNetworkProp(key, value, onChange)
-      // },
-      // control(options) {
-      //   // TODO: only allow on user interaction
-      //   // TODO: show UI with a button to release()
-      //   entity.control = world.input.bind({
-      //     ...options,
-      //     priority: 50,
-      //     object: entity,
-      //   })
-      //   return entity.control
-      // },
-      ...this.base.getProxy(),
+      control(options) {
+        // TODO: only allow on user interaction
+        // TODO: show UI with a button to release()
+        entity.control = world.controls.bind({
+          ...options,
+          priority: ControlPriorities.APP,
+          object: entity,
+        })
+        return entity.control
+      },
+      ...this.root.getProxy(),
     }
   }
 }