+++ CHANGELOG.md
@@ -13,6 +13,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [0.3.0]
+
+### Added
+- Environment variable to limit model upload size
+- Node.traverse(callback)
+- Ability to disable world saving completely
+- Initial app networking
+- Temporary skybox
+- Re-enabled stats via /stats chat command
+- Let players know when they are disconnected from the world
+- Add /health endpoint
+- Action node for interactive apps
+- Script to clean up orphaned blueprints/files in a world (npm run world:clean) [experimental]
+- Expose Collider.convex to script runtime
+- Expose LOD.insert to script runtime
+- Initial docs!
+
+### Changed
+- Use "geometry" type for Mesh and Collider nodes (instead of "custom")
+- Enter and Leave world events are now a player object instead of just the networkId
+
+### Fixed
+- Production source-maps issue
+- Errors using geometry with morph targets
+- Server crash attempting to load an asset that does not exist
+- World event unhandled errors
+
 ## [0.2.0] - 2025-01-14
 
 ### Added
@@ -53,6 +80,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic project structure
 - Core functionality from original project
 
-[Unreleased]: https://github.com/hyperfy-xyz/hyperfy/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/hyperfy-xyz/hyperfy/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/hyperfy-xyz/hyperfy/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/hyperfy-xyz/hyperfy/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/hyperfy-xyz/hyperfy/releases/tag/v0.1.0